Extract WCFM Marketplace client to its namespace
And move logging to APIClient. This low level thing doesn't belong to
a model and neither to these tests.
